+++ esp-hal/CHANGELOG.md
@@ -24,7 +24,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Allow DMA to/from psram for esp32s3 (#1827)
 - DMA buffers now don't require a static lifetime. Make sure to never `mem::forget` an in-progress DMA transfer (consider using `#[deny(clippy::mem_forget)]`) (#1837)
 - Peripherals (where possible) are now explicitly reset and enabled in their constructors (#1893)
-- Deprecated `Rtc::get_time_raw` since it doesn't respect boot time and therefore won't react to setting RTC time (#1883)
+- Renamed `Rtc::get_time_raw` to `Rtc::get_rtc_time_raw` (#1883)
 
 ### Fixed